Rework file hierarchy. Hoist product state to app level
