+++ src/components/store/Store.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import './Store.css';
-import Sidebar from '../sidebar/Sidebar';
-import Products from '../products/Products';
+import Sidebar from './sidebar/Sidebar';
+import Products from './products/Products';
 
 import { getProducts, getCategories } from '../../services/store-service';
 
@@ -13,8 +13,10 @@ class Store extends React.Component {
       products: [],
       categories: [],
       selectedCategory: null,
+      currentProductPage: 0,
     };
     this.onCategoryChange = this.onCategoryChange.bind(this);
+    this.onPageChange = this.onPageChange.bind(this);
   }
 
   // Fetch categories and products
@@ -34,21 +36,39 @@ class Store extends React.Component {
 
   onCategoryChange(category) {
     this.setState({
+      currentProductPage: 0,
       selectedCategory: category,
     });
   }
 
+  onPageChange(page) {
+    this.setState({
+      currentProductPage: page,
+    });
+  }
+
   render() {
-    const { products, categories, selectedCategory } = this.state;
+    const {
+      products, categories, selectedCategory, currentProductPage,
+    } = this.state;
 
     return (
       <main className="store">
         <aside className="sidebar">
-          <Sidebar categories={categories} onCategoryChange={this.onCategoryChange} selectedCategory={selectedCategory} />
+          <Sidebar
+            categories={categories}
+            onCategoryChange={this.onCategoryChange}
+            selectedCategory={selectedCategory}
+          />
         </aside>
 
         <section className="products" role="main">
-          <Products products={products} selectedCategory={selectedCategory} />
+          <Products
+            products={products}
+            selectedCategory={selectedCategory}
+            currentPage={currentProductPage}
+            onPageChange={this.onPageChange}
+          />
         </section>
       </main>
     );
