+++ src/components/store/products/paginator/Paginator.css
